+++ CHANGES.md
@@ -8,6 +8,7 @@
 * Fixed changes to the launcher icon not being applied after a self-update on Linux.
 * `cmd/validator`: fix only the first missing bundle URL being reported.
 * Fix missing `LauncherUpdate` in deployment-config leading to error when documentation explicitly allows it.
+* Fix false positives in log warning about program name on disk having diverged from what is configured in the embedded launcher-config on Windows and MacOS.
 ### Changes
 * trivrost now allows users to continue past failed self-updates as well as omitted changes to system bundles in [system mode](docs/lifecycle.md#system-mode) installations, unless [further configuration](docs/deployment-config.md) is made; see the new field `IsUpdateMandatory` in deployment-config.